pyproject.toml
[project]
name = "sp-lightrag-agent"
version = "0.1.0"
description = "Minimal QA agent for construction standards."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.9",
    "lightrag-hku>=1.4.9.11",
    "nicegui>=3.8.0",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]
